Fix object positions, add enemy mouse hover
I forgot that actors are anchored at their center, so setting the x/y to the coordinates coming in from Tiled caused them to be offset by half their height and width.

The hover color change is a proof of concept that I can detect and react to the player selecting an enemy path.