Track which cell characters are assigned to Build a grid of cells Add function to get neighbors of a given cell
This setup allows us to check whether any given character is a valid target or not via:
* Check if their cell is a neighbor of the tail cell of the current path (doesn't exist yet, using the player's cell for now) or the player if there is no path
* Check if the enemy occupant in that cell is the same type as the occupant of the tail cell (only if there is a path already)

Then we'll need to handle selecting an already-selected enemy in the path to move the path back, etc. But for now, the highlight color of the cell proves that the base concept works.