Add path building Consolidate logic to mostly be controlled by Cells
This allows you to build out the path you're going to attack and draws a line to show you the path you have selected so far. It disallows selecting other colors once you've chosen the first color, and allows reverting the path by clicking on an earlier selected enemy.

The logic for deselecting is if you click on the last enemy in the path (even if it's the only one), it deselects that enemy. If you click on an earlier one in the path, it deselects up to (but not including) that one as you are probably trying to reset and go a different direction from that enemy.

I think it makes sense...we'll see. I am also not completely happy about the control flow/handling of responsibilities along the way, but it works and I'll revisit if I have time.

Next up: moving through and killing the enemies you've selected. AKA: the fun part!