Add score ui
This took me down a huge rabbit hole of discovering that I was out of date on the core pieces (vite, excalibur, tiled plugin) and once I updated them I discovered that apparently v0.29 was significant and had a lot of changes. So...that took a minute to patch up.

The main problem here was that the tile background was coming in with a higher z-index than player and enemies so they were hidden.

But now there's a POC score ui, so...time...well spent?