Add score ui, take 2
Since the newer version of Excalibur screwed up pixel rendering something awful, I reverted back to the older version. This older version handled fonts differently (the Excalibur version that changed rendering also added a built-in font loader) so that needed to be adjusted.

I can't define canvasElementId on the engine for some reason because it's massively offsetting the pointer (when the pointer is in the bottom right corner, it thinks it's near the center) so I need to figure that out later if I want to be able to have a centered canvas. Maybe the pointer needs to be anchored differently?