Update to Excalibur 0.30
Note that the change ("exclude" entry) in vite.config.ts is load-bearing: there's a problem introduced by 0.30 where tile layers aren't processed correctly, and this fixes it for dev builds ("npm run dev"). This does *not* fix it for production builds, so I may end up having to revert this (or switch to Parcel) if that isn't fixed.

ref: https://discord.com/channels/1195771303215513671/1317152887469772810

The other big discovery is that somewhere between 0.27 and 0.30 actors are no longer able to render anything by just setting a Color. They will always draw that color and no amount of changing the Color will update what is drawn; you have to make your own Rectangle and adjust *its* color instead.

The Tiled assset changes are because I thought I could automatically instantiate a class by naming them, but that turned out to not work the way I wanted.