Add danger and death/lose condition
Some enemies will enrage sometimes and if you end your turn next to one, they'll damage you. If you take 3 damage you die and have to restart.

This also overhauls level-loading because I wasn't able to find an easy way to reload the level you were currently on, and switching back to an existing scene was just going back to the existing GameScene, so this spawns a new one every time we load into a level and establishes a flow of the intended sequence.