Try to fix the wrong-path problem
On several occasions on the first selection of a new turn I've seen clicking a path cause a line to be drawn to some other enemy first before the one that was clicked. The path array doesn't have old data in it, so the only thing that makes sense is the 'hovered' flag staying on for a cell/enemy when it shouldn't. I'm hoping this fixes it.