Use green enemies on the brown stage
Red was used for the original stage because of the green background, to make it easier to read. I felt like red had the same problem on the brown background and green looks better to me. It's a little close to the blue color to my eye, but the shapes still help keep them distinct.