Play audio when enemies enrage
This also delays subsequent enragings when multiple enemies enrage at once so that the sounds don't all play at once.

I also saw a log that said it enraged 3 enemies but only 2 were actually angry in the game, so I'm convinced it somehow picked a dead enemy to enrage. I don't know how that would be possible, but I added a check for it anyway.